Added Python3 installation steps
Not sure if anything else is needed for the landsat-util/downloader.py script
homura installation in a virtualenv may be problematic. An alternative could be sudo apt-get install python3-homura.

Hopefully, we'll get around to including landsat-util/downloader.py as a module in this repository and the lasero repository. This will remove the extra homura installation step.
We should also include a script to download the supplementary material directly and automatically extract the text file to the working directory with cleaner names (spaces in path removed)
We should also write a script to extract the scene ids and return them as a list from the file.